Recursively decapsulate the packet.
Extensively tested by watching youtube while running sniff. Also,
highly entertaining.

The decapsulate function still isn't quite right yet, but works.
It will return a list of records and the binary as a payload. If
any of the headers has a problem, it will return {unsupported, Data}
or {truncated, Data} (which will crash sniff).

decapsulate/2 will return stop when it's reached the end of the
headers.  The rationale is that we will need to support parsing
other headers for IPv6 (e.g., the IPv6 Hop-by-Hop Options header)
and eventually, decoding application headers should be supported.
Maybe by allowing the caller to pass in a fun.

Remove the jumbo payload check, until we support the ipv6 headers.