Disallow use of file option with setuid epcap
If the "file" option is used with a setuid binary potentially any file
could be read by epcap. Modify epcap to error and exit in this case.

epcap will not exit if sudo is used since sudo sets uid and euid
to 0.  This can be worked around by hardcoding the arguments to epcap
in sudoers.conf.