API change: return {pkthdr, list()}
Somehow the second element of the tagged pkthdr tuple was a tuple rather
than a proplist. I blame my keyboard: must've typed a { instead of a [ ...
and a ei_x_encode_tuple_header instead of ei_x_encode_list_header() <.<

This is a backwards incompatible change. The idea was that the tuple was
like the pkthdr struct and so was ordered and tagged with the field name.
The list is still ordered but can be used as a proplist.