Use a short timeout for pcap_next_ex()
The behaviour of the timeout passed to pcap differs between BSD and
Linux. On Linux, a timeout of 0 means "return after the next packet is
received". On BSD, it means "return after the packet buffer is full".

Revert to the old timeout of 500 ms.

Thanks to Elliot Crosby-McCullough for discovering and debugging
the problem on Mac OS X!