epcap: enable support for injecting packets
Allow epcap to generate packets from the network device using
pcap_inject(3PCAP). By default, support is compiled in but disabled. To
enable injection, start epcap with the {inject, true} option.

If send is disabled, data is discarded by the port.

If send is enabled, errors returned by pcap_inject are fatal. The port
will terminate, crashing the gen_server. Partial writes are not
considered to be errors and are ignored.

This patch changes the behaviour of the gen_server/port interaction
slightly: previously any activity on the port's stdin would cause the
port to exit. Now the port will exit on error (from read or pcap_inject)
or on EOF.