test: add a test for sniffing a packet
Simple test for retrieving a packet. Note, there is a timing issue here.
Even though the port has been started, epcap is not monitoring the
interface yet. If the test does a TCP connection immediately with a
timeout of 100, epcap does not see the SYN.

Increasing the timeout to 2000 gives epcap enough time to being
monitoring the interface.

This test brings up two possible issues with epcap:

1. should epcap send a message to the process when it is ready?

2. Should the message sent back from epcap include an identifier?

    {packet, ID, DataLinkType, Time, Length, Packet}

Where ID is the pid of the gen_server (or it could be the port() or the
OS pid).