Suppress compile warning on Solaris
And this is why we can't have nice things, Solaris. Solaris still uses
caddr_t (which is an alias for char*) for iov_base. Apparently this
has been changed to void* in Solaris 11 but on the Solaris variant I am
testing on (SmartOS, whose uname humourously reports itself as 5.11),
the default for iov_base is caddr_t unless you ask for XPG4_2 which
will include a whole world of pain when you just wanted to suppress a
compiler warning without the evils of casting ...

So the alternatives seem to be ifdef'ing the code where iov_base is set
just for Solaris, adding a EPCAP_IOV_BASE macro where a cast to caddr_t
is done just for Solaris or punishing everybody by muddying the code
with a cast to void with all the oddities and potentials for undefined
behaviour that entails but, hey, less typing!