supervisor: forward signals to subprocess
The epcap pcap subprocess may not exit if the epcap supervisor unix
process is killed. Trap and forward signals to the subprocess.

If the seccomp sandbox was enabled for the supervisor process, violation
of the policy resulted in the process being sent an untrappable
SIGSYS. Modify the policy to send a catchable SIGSYS.

On FreeBSD, the signal handler is not strictly necessary. Process
cleanup works reliably with capability mode: when the epcap supervisor
process exits, the process descriptor is closed and the subprocess sent
a SIGKILL. Signals (except for SIGKILL) sent from outside capability
mode are ignored.