env: add option to set environment variable
Add a new option to set an environment variable in the port process:

    # equivalent to: {cluster_id, 1}
    epcap:start([{env, "PCAP_PF_RING_CLUSTER_ID=1"}])

libpcap uses environment variables for configuration.

Alternatively, the environment can be set in the shell environment and
{exec, "sudo -E"} used.

The advantage of using a dedicated option (cluster_id vs env) is that,
theoretically, given an appropriate typespec, dialyzer will warn if the
option is misspelled (as was done with `cluster_id` in the test!).