Add 'immediate' option
Currently the immediate mode cannot be selected explicitely. On some
platform the mode is enabled by default, on others (Linux) it is not.
If immediate mode is enabled, the timeout option will be ignored.

This causes the test suite to fail due to timeouts on Linux with
libpcap >1.5.0.

Enable immediate mode by default on all platforms.

Add a '-I' option to epcap to enable/disable the immediate mode.
Expose this via the boolean 'immediate' option.

Note that the 'timeout' option is only honoured if the immediate mode is
disabled, so using {timeout, To_ms} makes only sense when also using
{immediate, false}.