Use errstrp to detect strtonum errors
Currently errno is used to check whether strtonum has encountered an
error. As it is common with errno, that variable is only set by
strtonum on errors and remains unmodified on success. Since the
error condition cannot be derived from the return value, a previously
set errno can lead to an error exit even if the string was a valid
number and within the range. The initial errno value is not specific,
so epcap may consequently fail on some platforms.

Check the errstrp output parameter instead to check whether an error has
appeared.